massive refactor, generalize frontend/backend api
