+++ qiling/debugger/qdb/arch/__init__.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python3
+#
+# Cross Platform and Multi Architecture Advanced Binary Emulation Framework
+#
+
+from .arch_x86 import ArchX86
+from .arch_mips import ArchMIPS
+from .arch_arm import ArchARM, ArchCORTEX_M
+
+if __name__ == "__main__":
+    pass
