+++ qiling/debugger/qdb/arch/arch_x86.py
@@ -1,0 +1,43 @@
+#!/usr/bin/env python3
+#
+# Cross Platform and Multi Architecture Advanced Binary Emulation Framework
+#
+
+from __future__ import annotations
+from typing import Mapping
+
+from .arch import Arch
+
+class ArchX86(Arch):
+    def __init__(self: ArchX86):
+        self.regs = (
+                "eax", "ebx", "ecx", "edx",
+                "esp", "ebp", "esi", "edi",
+                "eip", "ss", "cs", "ds", "es",
+                "fs", "gs", "ef",
+                )
+
+    def read_insn(self, address: int) -> bytes:
+        # due to the variadic lengh of x86 instructions ( 1~15 )
+        # always assume the maxium size for disassembler to tell
+        # what is it exactly.
+
+        return self.read_mem(address, 15)
+
+    @staticmethod
+    def get_flags(bits: int) -> Mapping[str, bool]:
+        """
+        get flags from ql.reg.ef
+        """
+
+        return {
+                "CF" : bits & 0x0001 != 0, # CF, carry flag
+                "PF" : bits & 0x0004 != 0, # PF, parity flag
+                "AF" : bits & 0x0010 != 0, # AF, adjust flag
+                "ZF" : bits & 0x0040 != 0, # ZF, zero flag
+                "SF" : bits & 0x0080 != 0, # SF, sign flag
+                "OF" : bits & 0x0800 != 0, # OF, overflow flag
+                }
+
+if __name__ == "__main__":
+    pass
