+++ qiling/debugger/qdb/branch_predictor/__init__.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python3
+#
+# Cross Platform and Multi Architecture Advanced Binary Emulation Framework
+#
+
+from .branch_predictor_x86 import BranchPredictorX86
+from .branch_predictor_mips import BranchPredictorMIPS
+from .branch_predictor_arm import BranchPredictorARM, BranchPredictorCORTEX_M
+
+if __name__ == "__main__":
+    pass
